hideaway the grungy details of configs_tree_view_data...
